requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain==0.3.25
langchain-openai==0.3.35
langchain-core==0.3.80
langgraph>=0.2.0
pydantic>=2.9.0
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.0.0

